Merges cached_property and cached_node_property; adds mixins for default no cache and default disk cache nodes.
The cached_property now takes a keyword argument use_cache_ctrl. It can also be used with no argument.
NoCacheMixin makes a node have no cache stores by default. DiskCacheMixin appends a disk cache, if necessary, to the cache stores by default.
