+++ podpac/core/utils.py
@@ -331,39 +331,160 @@ def _get_from_url(url):
     return r.text
 
 
-def cached_property(fn):
-    """Decorator that creates a property that is cached as a private attribute."""
+def cache_func(key, depends=None):
+    """
+    Decorating for caching a function's output based on a key.
 
-    key = "_podpac_cached_property_%s" % fn.__name__
-
-    @property
-    def wrapper(self):
-        if hasattr(self, key):
-            value = getattr(self, key)
-        else:
-            value = fn(self)
-            setattr(self, key, value)
-        return value
-
-    return wrapper
+    Parameters
+    -----------
+    key: str
+        Key used for caching.
+    depends: str, list, traitlets.All (optional)
+        Default is None. Any traits that the cached property depends on. The cached function may NOT
+        change the value of any of these dependencies (this will result in a RecursionError)
 
 
-def cached_node_property(fn):
-    """Decorator that creates a property that is cached as a private attribute and in the Node cache_ctrl."""
+    Notes
+    -----
+    This decorator cannot handle function input parameters.
 
-    key = "_podpac_cached_property_%s" % fn.__name__
+    If the function uses any tagged attributes, these will essentially operate like dependencies
+    because the cache key changes based on the node definition, which is affected by tagged attributes.
 
-    @property
-    def wrapper(self):
-        if hasattr(self, key):
-            value = getattr(self, key)
-        elif self.has_cache(key):
-            value = self.get_cache(key)
-            setattr(self, key, value)
-        else:
-            value = fn(self)
-            setattr(self, key, value)
-            self.put_cache(value, key)
-        return value
+    Examples
+    ----------
+    >>> from podpac import Node
+    >>> from podpac.core.node import cache_func
+    >>> import traitlets as tl
+    >>> class MyClass(Node):
+           value = tl.Int(0)
+           @cache_func('add')
+           def add_value(self):
+               self.value += 1
+               return self.value
+           @cache_func('square', depends='value')
+           def square_value_depends(self):
+               return self.value
 
-    return wrapper
+    >>> n = MyClass(cache_ctrl=None)
+    >>> n.add_value()  # The function as defined is called
+    1
+    >>> n.add_value()  # The function as defined is called again, since we have specified no caching
+    2
+    >>> n.cache_ctrl = CacheCtrl([RamCacheStore()])
+    >>> n.add_value()  # The function as defined is called again, and the value is stored in memory
+    3
+    >>> n.add_value()  # The value is retrieved from disk, note the change in n.value is not captured
+    3
+    >>> n.square_value_depends()  # The function as defined is called, and the value is stored in memory
+    16
+    >>> n.square_value_depends()  # The value is retrieved from memory
+    16
+    >>> n.value += 1
+    >>> n.square_value_depends()  # The function as defined is called, and the value is stored in memory. Note the change in n.value is captured.
+    25
+    """
+    # This is the actual decorator which will be evaluated and returns the wrapped function
+    def cache_decorator(func):
+        # This is the initial wrapper that sets up the observations
+        @functools.wraps(func)
+        def cache_wrapper(self):
+            # This is the function that updates the cached based on observed traits
+            def cache_updator(change):
+                # print("Updating value on self:", id(self))
+                out = func(self)
+                self.put_cache(out, key, overwrite=True)
+
+            if depends:
+                # This sets up the observer on the dependent traits
+                # print ("setting up observer on self: ", id(self))
+                self.observe(cache_updator, depends)
+                # Since attributes could change on instantiation, anything we previously
+                # stored is likely out of date. So, force and update to the cache.
+                cache_updator(None)
+
+            # This is the final wrapper the continues to fetch data from cache
+            # after the observer has been set up.
+            @functools.wraps(func)
+            def cached_function():
+                try:
+                    out = self.get_cache(key)
+                except podpac.NodeException:
+                    out = func(self)
+                    self.put_cache(out, key)
+                return out
+
+            # Since this is the first time the function is run, set the new wrapper
+            # on the class instance so that the current function won't be called again
+            # (which would set up an additional observer)
+            setattr(self, func.__name__, cached_function)
+
+            # Return the value on the first run
+            return cached_function()
+
+        return cache_wrapper
+
+    return cache_decorator
+
+
+def cached_property(*args, **kwargs):
+    """
+    Decorator that creates a property that is cached.
+
+    Keyword Arguments
+    -----------------
+    use_cache_ctrl : bool
+        If True, the property is cached using the Node cache_ctrl. If False, the property is only cached as a private
+        attribute. Default False.
+
+    Examples
+    --------
+
+    >>> class MyNode(Node):
+            # property that is recomputed every time
+            @property
+            def my_property(self):
+                return 0
+
+            # property is computed once for each object
+            @cached_property
+            def my_cached_property(self):
+                return 1
+
+            # property that is computed once and can be reused by other Nodes or sessions, depending on the cache_ctrl
+            @cached_property(use_cache_ctrl=True)
+            def my_persistent_cached_property(self):
+                return 2
+    """
+
+    use_cache_ctrl = kwargs.pop("use_cache_ctrl", False)
+
+    if args and (len(args) != 1 or not callable(args[0])):
+        raise TypeError("cached_property decorator does not accept any positional arguments")
+
+    if kwargs:
+        raise TypeError("cached_property decorator does not accept keyword argument '%s'" % list(kwargs.keys())[0])
+
+    def d(fn):
+        key = "_podpac_cached_property_%s" % fn.__name__
+
+        @property
+        def wrapper(self):
+            if hasattr(self, key):
+                value = getattr(self, key)
+            elif use_cache_ctrl and self.has_cache(key):
+                value = self.get_cache(key)
+                setattr(self, key, value)
+            else:
+                value = fn(self)
+                setattr(self, key, value)
+                if use_cache_ctrl:
+                    self.put_cache(value, key)
+            return value
+
+        return wrapper
+
+    if args:
+        return d(args[0])
+    else:
+        return d
