+++ podpac/utils.py
@@ -2,8 +2,9 @@
 Utils Public Module
 """
 
-# REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
+# REMINDER: update api docs (doc/source/api.rst) to reflect changes to this file
 
 
-from podpac.core.utils import create_logfile
+from podpac.core.utils import create_logfile, cached_property, cache_func
 from podpac.core.cache import clear_cache
+from podpac.core.node import NoCacheMixin, DiskCacheMixin, S3Mixin
